+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## unreleased
+## 1.15.0
 
 ### Features